requirements.txt
numpy>=1.21.0
opencv-python>=4.8.0
matplotlib>=3.7.1
gradio>=4.0.0
Pillow>=9.0.0
ultralytics>=8.0.0
tqdm>=4.65.0
scikit-learn>=1.2.2
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.1 